feat(trajectory_validator): rename autoware_trajectory_safety_filter to autoware_trajectory_validator (#12312)
* fix: rename safety filter to validator


* Renaming additional artifacts


* fix: build error


* fix: rename config


* fix: parameter namespace


* fix: use safety namespace


---------
