--- planning/autoware_trajectory_safety_filter/package.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0"?>
-<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
-<package format="3">
-  <name>autoware_trajectory_safety_filter</name>
-  <version>0.50.0</version>
-  <description>The autoware_trajectory_safety_filter package</description>
-
-  <maintainer email="alqudah.mohammad@tier4.jp">Alqudah Mohammad</maintainer>
-  <maintainer email="daniel.sanchez@tier4.jp">Daniel Sanchez</maintainer>
-  <maintainer email="go.sakayori@tier4.jp">Go Sakayori</maintainer>
-  <maintainer email="maxime.clement@tier4.jp">Maxime Clement</maintainer>
-  <maintainer email="Kotakkucu@gmail.com">Takumi Odashima</maintainer>
-  <maintainer email="yukihiro.saito@tier4.jp">Yukihiro Saito</maintainer>
-  <maintainer email="zulfaqar.azmi@tier4.jp">Zulfaqar Azmi</maintainer>
-
-  <license>Apache License 2.0</license>
-
-  <author email="go.sakayori@tier4.jp">Go Sakayori</author>
-
-  <buildtool_depend>ament_cmake</buildtool_depend>
-  <buildtool_depend>autoware_cmake</buildtool_depend>
-
-  <depend>autoware_boundary_departure_checker</depend>
-  <depend>autoware_internal_planning_msgs</depend>
-  <depend>autoware_lanelet2_utils</depend>
-  <depend>autoware_map_msgs</depend>
-  <depend>autoware_motion_utils</depend>
-  <depend>autoware_planning_msgs</depend>
-  <depend>autoware_universe_utils</depend>
-  <depend>autoware_utils_debug</depend>
-  <depend>autoware_utils_geometry</depend>
-  <depend>autoware_utils_rclcpp</depend>
-  <depend>autoware_vehicle_info_utils</depend>
-  <depend>generate_parameter_library</depend>
-  <depend>geometry_msgs</depend>
-  <depend>nav_msgs</depend>
-  <depend>pluginlib</depend>
-  <depend>rclcpp</depend>
-  <depend>rclcpp_components</depend>
-  <depend>tl_expected</depend>
-
-  <test_depend>ament_cmake_ros</test_depend>
-  <test_depend>ament_lint_auto</test_depend>
-  <test_depend>autoware_lint_common</test_depend>
-  <test_depend>autoware_test_utils</test_depend>
-
-  <export>
-    <build_type>ament_cmake</build_type>
-  </export>
-</package>
