--- planning/autoware_trajectory_safety_filter/test/dummy_plugin/dummy_plugins.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0"?>
-<library path="dummy_safety_filter_plugin">
-  <class type="autoware::trajectory_safety_filter::plugin::DummyFilter"
-         base_class_type="autoware::trajectory_safety_filter::plugin::SafetyFilterInterface">
-    <description>A predictable dummy filter for testing the main node</description>
-  </class>
-</library>
