+++ planning/autoware_trajectory_validator/launch/trajectory_validator.launch.xml
@@ -1,0 +1,20 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<launch>
+  <arg name="trajectory_validator_param_path" default="$(find-pkg-share autoware_trajectory_validator)/config/trajectory_validator.param.yaml"/>
+  <arg name="input_trajectories" default="~/input/trajectories"/>
+  <arg name="output_trajectories" default="~/output/trajectories"/>
+  <arg name="input_vector_map" default="~/input/vector_map"/>
+  <arg name="input_odometry" default="~/input/odometry"/>
+  <arg name="input_objects" default="~/input/objects"/>
+  <arg name="input_acceleration" default="~/input/acceleration"/>
+
+  <node pkg="autoware_trajectory_validator" exec="autoware_trajectory_validator_node" name="trajectory_validator_node" output="screen">
+    <param from="$(var trajectory_validator_param_path)" allow_substs="true"/>
+    <remap from="~/input/trajectories" to="$(var input_trajectories)"/>
+    <remap from="~/output/trajectories" to="$(var output_trajectories)"/>
+    <remap from="~/input/lanelet2_map" to="$(var input_vector_map)"/>
+    <remap from="~/input/odometry" to="$(var input_odometry)"/>
+    <remap from="~/input/objects" to="$(var input_objects)"/>
+    <remap from="~/input/acceleration" to="$(var input_acceleration)"/>
+  </node>
+</launch>
