+++ planning/autoware_trajectory_validator/plugins.xml
@@ -1,0 +1,21 @@
+<?xml version="1.0"?>
+<library path="autoware_trajectory_validator_plugins">
+  <class type="safety::OutOfLaneFilter"
+    base_class_type="autoware::trajectory_validator::plugin::ValidatorInterface">
+    <description>
+      Filters trajectories that go out of lane boundaries
+    </description>
+  </class>
+  <class type="safety::CollisionFilter"
+    base_class_type="autoware::trajectory_validator::plugin::ValidatorInterface">
+    <description>
+      Filters trajectories based on Time To Collision (TTC) with predicted objects
+    </description>
+  </class>
+  <class type="safety::VehicleConstraintFilter"
+    base_class_type="autoware::trajectory_validator::plugin::ValidatorInterface">
+    <description>
+      Filters trajectories that violate vehicle constraints such as maximum speed, acceleration, and steering angle
+    </description>
+  </class>
+</library>
