+++ planning/autoware_trajectory_validator/trajectory_validator.param.yaml
@@ -1,0 +1,27 @@
+/**:
+  ros__parameters:
+    filter_names:
+      [
+        "autoware::trajectory_validator::plugin::OutOfLaneFilter",
+        "autoware::trajectory_validator::plugin::CollisionFilter",
+        "autoware::trajectory_validator::plugin::VehicleConstraintFilter",
+      ]
+    # available filters:
+    #  "autoware::trajectory_validator::plugin::OutOfLaneFilter"
+    #  "autoware::trajectory_validator::plugin::CollisionFilter"
+    #  "autoware::trajectory_validator::plugin::VehicleConstraintFilter"
+
+    out_of_lane:
+      time: 2.0
+      min_value: 0.0
+
+    collision:
+      time: 3.0
+      min_value: 2.0
+
+    vehicle_constraint:
+      max_speed: 16.7 # m/s
+      max_acceleration: 5.0 # m/s^2
+      max_deceleration: 5.0 # m/s^2 (positive but represents deceleration)
+      max_steering_angle: 0.8 # rad
+      max_steering_rate: 0.3 # rad/s
